refactor: eventType enum 객체 outbox 패키지로 이동
