+++ src/main/java/com/gloddy/server/article/application/ArticlePayloadService.java
@@ -4,7 +4,6 @@ import com.gloddy.server.article.domain.Article;
 import com.gloddy.server.article.domain.dto.ArticlePayload;
 import com.gloddy.server.article.domain.dto.in.GroupArticlePayloadEventType;
 import com.gloddy.server.article.domain.handler.ArticleQueryHandler;
-import com.gloddy.server.messaging.adapter.group.event.GroupArticleEventType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
